:tractor: :newspaper: [IMRF-83] Split Importify.Resolution into modules
